添加上传文件、下载文件方法

+++ commontoolslibrary/src/main/java/com/qishui/commontoolslibrary/http/callback/GsonCallBack.java
@@ -90,7 +90,7 @@ public abstract class GsonCallBack<T> implements ICallBack {
     }
 
     @Override
-    public void inProgress(int progress) {
+    public void inProgress(float progress) {
 
     }
 }
